Removed some dubious filetyping logic.
On saving a file if the type was text no filetype information was
saved and instead it relied on the default for unknown filetypes being
text.
The default unknown filetype is now &FE4,plus the above logic would
have conflicted with the MimeMap code if for example a file of type
text was saved but which had an extension which was in the mimemap of
type something other than text.

Version 2.24. Tagged as 'LanManFS-2_24'