Ammendments to short name translation table
The table was ambiguously translating () to () and also [] to (). Now, we error names containing ,[] in the same manner as DOSFS.
The substitutions for ^% have also been changed to match DOSFS so copying from one to the other doesn't give two representations if viewed on a PC.
Marked both tables as const to save 512B of RMA.
This change is slightly academic since most of the time the long filename resolution is in use.

Tested against WinXP SP3 with a build with long filename support turned off.

Version 2.45. Tagged as 'LanManFS-2_45'