Add a quick exit for the timeout=0 case
Careful analysis of packet captures by Colin Granville reveals that LanManFS waits for at least 1cs when no timeout was requested, so add a quick exit for that case. Timeouts of >= 1cs still delay for at least the requested time, and up to 0.999cs longer as before.
This reduces the turn around time which is most beneficial for interfaces faster than 10baseT.

Version 2.48. Tagged as 'LanManFS-2_48'