Remove obsolete NetBEUI transport from build.
Remove the obsolete NetBIOS frames (NetBEUI) transport from the build.
I haven't deleted "c/NetBIOS" yet, in case it's useful as a reference,
but I wrapped the code with "#if 0" to ensure that it won't be built,
even if it were to be added back to the Makefile by mistake.