Fix SMB_Negotiate timeout when SMB1 is disabled.
SMB servers that have disabled SMB1 will return a dialect index of
0xFFFF in the SMB_COM_NEGOTIATE response, which caused a 40 second
timeout because it wasn't recognized as a negotiation failure.

Return EPROTOCOLERR from SMB_Negotiate() for this case so that we
fail immediately instead of spinning the hourglass for 40 seconds.
The SMB 2.x feature will add new dialects to the negotiation request.