Add a comment and debug printf.
Add a code comment for the error called ERRnoaccess in the spec
and ENOACCESS in the LanMan code, since it's a common error return.

Add a debug printf for read timeouts and other errors that occur
while reading an SMB response from the socket. Otherwise, it's
impossible to tell from the debug log what type of command failed.