Fix file r/w timeout on Vista and newer.
Test the correct CAP_RAW_MODE bit in the SMB 1 negotiation response,
fixing a timeout when reading or writing on Windows Vista, Server 2008,
and newer, which immediately close the socket if you try to use the
"raw" read or write SMB 1 commands.

We should be using SMB 2 on Vista and newer, so there's no need
to optimize SMB 1 performance, but it should at least work well
enough for comparison and benchmarking purposes. Success with the
SMB commands that transfer 4K at a time is much better than failure
with the no-longer-supported-by-Windows raw read/write commands.

With this commit, SMB 1 now works well enough that I can start writing
the new SMB 2 code paths, which should be faster, more maintainable,
more secure, and more fun to work with than the horrible SMB 1.