Remove SMB1 raw r/w code paths.
The READ_RAW and WRITE_RAW commands aren't supported by Windows 2000
or newer, or Samba, so let's #ifdef out the code paths for them.
Attempting to use the raw commands results in either incorrect data
(Samba 3.6) or the server closes the connection (Vista or newer),
so there's no reason to keep this untestable code in the module.