Undo bad commit to CAP_RAW_MODE check; add debug.
The original test for CAP_RAW_MODE was correct; I'd mixed up words
and bytes. This means there's something broken in our READ_RAW and
WRITE_RAW commands, or in Samba 3.6's handling of them.

It's better to leave these commands disabled to focus on adding SMB2.
The MS-SMB spec says in Appendix A, note 29, that Windows Server 2008
and later (and Vista, as I discovered) don't support SMB_COM_READ_RAW
or SMB_COM_WRITE_RAW and disconnect the client by closing the socket
if either command is received from the client.

That's why I mistakenly assumed that the test was wrong. In reality,
Microsoft is advertising a feature that they don't support and then
closing the connection if a client tries to use that feature!