Comment out SMB1 anti-idle logic.
The timer callback to send no-op SMB1 commands to connected
servers to try to keep buggy old Windows servers from closing
the socket is not compatible with SMB2, and it's also buggy
if any SMB1 request is active that bypasses Do_SMB().

The only SMB1 commands that bypass Do_SMB() are SMB_ReadRaw()
and SMB_WriteRaw(), which I've #ifdef'd out due to unrelated bugs.
The main purpose of this commit is to prepare for SMB2 by removing
irrelevant code paths, including the timer callback itself.