Start replacing SMB1 code with new SMBX client.
Removed most of the original SMB protocol handling code, keeping
the layer of code that interfaces with FileSwitch and the Omni client.
Imported most of the source files for the new SMBX client in order to
continue the porting process by working on individual source files.

The new code has additional license requirements, namely that the
original authors ask to be credited somewhere in the open source
acknowledgements file distributed with binaries of the software.
Current RISC OS binary distributions should already have one or
more software license acknowledgements files that can be added to.

Added a new "doc/SMBX_Changes.md" describing the SMB 2/3 project goals,
the new license requirements, system variables and other tuneable options,
and the names of the imported files, major data structures, and details
of memory management, error handling, RISC OS / UTF-16 two-way filename
conversion, and other aspects of this port.

This version of the code doesn't compile or link, except for the
source files in the current OBJS target in the Makefile. Additional
source files will be restored to the build as they're updated.

The purpose of this commit is to checkpoint the source tree with
the new header files, memory management, and error mapping from
NTSTATUS and POSIX errno values to FileSwitch-compatible errors
with localisable error strings.

Apple's corecrypto libraries aren't available under a suitable
open-source license, so replacements from LibreSSL 3.7.3 (latest
stable release) will be substituted in a later commit for use by
the encryption and signing code in "c/smb_crypt". The LibreSSL code
license requirements have already been added to the docs.